% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASLQC.r
\name{asl.QC}
\alias{asl.QC}
\title{aslprep.QC}
\usage{
asl.QC(filename = "ASLPREP_QC.csv")
}
\arguments{
\item{filename}{Filename of the .csv file. Set to \code{ASLPREP_QC.csv} by default}
}
\value{
outputs a .csv file containing the aslprep QC data
}
\description{
collating all aslprep QC data into a single .csv file
}
\details{
This function collates all aslprep QC data into a single .csv file
}
\examples{
\dontrun{
aslprep.QC()
}
}
