% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIFTI_subj_avg.R
\name{CIFTI_subj_avg}
\alias{CIFTI_subj_avg}
\title{CIFTI_subj_avg}
\usage{
CIFTI_subj_avg(rds, filename)
}
\arguments{
\item{rds}{The filepath of the \code{FSLRvextract}-generated .rds file}

\item{filename}{The filename of the output subject-averaged .rds file}
}
\value{
outputs a .rds file containing a matrix where each row represent each subject and each column represent each vertex.
}
\description{
Averaging multiple runs of vertex-wise(CIFTI) data within subject
}
\details{
Some fMRI datasets contain multiple runs per subject. Currently XCP-d generates ReHo or ALFF data for these multiple runs without concatenating the runs within a subject. This function averages the vertex-wise data such that each subject will only have one row of cifti data.
}
\examples{
\dontrun{
CIFTI_subj_avg(rds = "COBRE_reho.rds", filename = "averaged/COBRE_reho.rds")
}
}
