% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIFTItoFC.R
\name{CIFTItoFC}
\alias{CIFTItoFC}
\title{CIFTItoFC}
\usage{
CIFTItoFC(
  path = "./",
  wb_path = "/home/junhong.yu/workbench/bin_rh_linux64",
  dtseries = "_space-fsLR_den-91k_desc-denoised_bold.dtseries.nii",
  timeseries = F,
  concat_subj = TRUE,
  atlas = 200,
  round,
  filename = "FC.rds"
)
}
\arguments{
\item{path}{The filepath to directory containing the subject folders. Set to \code{./.} by default.}

\item{wb_path}{The filepath to the workbench folder that you have previously downloaded and unzipped. Set to \verb{/home/junhong.yu/workbench/bin_rh_linux64} by default}

\item{dtseries}{The filename extension of the fMRI volumes. Set to \verb{_space-fsLR_den-91k_desc-denoised_bold.dtseries.nii} by default}

\item{timeseries}{If set to \code{TRUE}, FC matrices will not be computed, instead the parcellated timeseries will be return in a list format, where each subject's parcellated timeseries data is contained within a list element. Set to \code{FALSE} by default.}

\item{concat_subj}{When set to \code{TRUE} (default), timeseries data from multiple runs/sessions of the same subject is first Z-scaled and then concatenated into a single larger timeseries data frame before computing the FC.}

\item{atlas}{The version (number of parcels) of the Schaefer atlas; it has to be in multiples of 100, from 100 to 1000. set to \code{200} by default. The specified atlas template will be automatically downloaded from here if it does not exist in the current directory.}

\item{round}{Number of decimal places to round the data to. Fewer decimal places require less disk space}

\item{filename}{Filename of the concatenated FC vector file. Set to \code{FC.rds} by default}
}
\value{
outputs N (number of subjects) x E (number of unique) matrices as a .rds file
}
\description{
Extracting FC matrices from CIFTI files
}
\details{
This function extracts FC matrices from CIFTI volumes postprocessed with XCP_d
}
\examples{
\dontrun{
CIFTItoFC(wb_path="bin_windows64/",dtseries="_ses-0._task-rest_acq-.._run-0._space-fsLR_den-91k_desc-denoised_bold.dtseries.nii",atlas = 200,filename="test.RDS")
}
}
