% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmriprep_func.R
\name{extract_headmotion}
\alias{extract_headmotion}
\title{extract_headmotion}
\usage{
extract_headmotion(filename = "motiondat.csv", start = 1)
}
\arguments{
\item{filename}{the desired filename, with a *.csv extension, of the output.Set to 'motiondat.csv' by default}

\item{start}{calculate motion parameters from this frame index. set to 1 by default}
}
\value{
outputs a .csv file containing columns of
\itemize{
\item \code{subj} Subject ID.
\item \code{no.frames} Number of frames.
\item \code{FD} Framewise displacement.
\item \code{RMSD} Root Mean Squared Displacement.
\item \code{FD.20} Number of frames with FD>0.2.
\item \code{RMSD.25} Number of frames with RMSD>0.25.
}
}
\description{
Extracting head motion measurements.
}
\details{
This function extracts head motion measurements from an fMRIprep output directory, and outputs these measurements in a .csv file
}
\examples{
\dontrun{
extract_headmotion()
}
}
