% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCP_dat_func.R
\name{extract_linksXCP}
\alias{extract_linksXCP}
\title{extract_linksXCP}
\usage{
extract_linksXCP(
  manifest = "datastructure_manifest.txt",
  task,
  surf,
  filename = "downloadlist.txt",
  subjects
)
}
\arguments{
\item{manifest}{the filepath of the manifest file. Set to \code{datastructure_manifest.txt} by default.}

\item{task}{name (case-sensitive) of the task, without numbers or directional labels. E.g., \code{REST}}

\item{surf}{name (case-sensitive) of the FreeSurfer surface data. E.g., \code{thickness}}

\item{filename}{The desired filename, with a *.txt extension, of the text file containing the links. Set to \code{download_list.txt} by default}

\item{subjects}{If specified the only download links from these subjects will be included}
}
\value{
outputs a .txt file containing the links filtered from the data manifest file
}
\description{
Extracting links of XCP-d minimal inputs and FreeSurfer surface data from data manifests
}
\details{
This function extracts links of XCP-d minimal inputs files and FreeSurfer surface data from NDA formatted data manifest files
}
\examples{
\dontrun{
extract_linksXCP(manifest = "datastructure_manifest.txt", task="REST",freesurfer = "thickness",subjects = c("0891566","0571649"))
}
}
