% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFSROI.R
\name{extractFS.ROI}
\alias{extractFS.ROI}
\title{extractFS.ROI}
\usage{
extractFS.ROI(basename = "dat")
}
\arguments{
\item{basename}{The base name of the .csv file. Set to \code{dat} by default.}
}
\value{
Returns outputs two .csv files
}
\description{
function to extract Destrieux atlas and ASEG ROIs
}
\details{
This function reads the *.table freesurfer output files and outputs two .csv files for the Destrieux and ASEG ROIs
}
\examples{
\dontrun{
extractFS.ROI(basename="dat")
}
}
