% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_curl_sh.R
\name{gen_curl_sh}
\alias{gen_curl_sh}
\title{gen_curl_sh}
\usage{
gen_curl_sh(
  sh,
  include_and,
  include_or,
  exclude,
  subjects,
  filename = "curl_cmd.sh"
)
}
\arguments{
\item{sh}{The filepath to sh file containing the curls commands as generated by openneuro}

\item{include_and}{The curl commands have to contain all of the keywords specfied here will be included}

\item{include_or}{The curl commands have to contain at least one of the keywords specfied here will be included}

\item{exclude}{The curl commands which contain at least one of the keywords specfied here will be excluded}

\item{subjects}{If specified only curl commands that include these subjects will be included}

\item{filename}{filename of the .sh output file. Set to \code{curl_cmd.sh} by default.}
}
\value{
a vector of curl commands
}
\description{
function to generate curl shellscript for downloading data from openneuro
}
\details{
This function searches the \code{sh} input file for keywords matching strings specified in \code{include_or}, \code{include_and}, \code{exclude} and \code{subjects}, and returns a vector of curl commands as well as output a .sh file
}
