% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headmotion_fmriprep.R
\name{headmotion.fmriprep}
\alias{headmotion.fmriprep}
\title{headmotion.fmriprep}
\usage{
headmotion.fmriprep(
  path = Sys.glob("sub-*"),
  filename = "rmsd.csv",
  threshold = 0.25
)
}
\arguments{
\item{filename}{Filename of the output.csv file. Set to \code{FD.csv} by default}

\item{threshold}{mean RMSD threshold for identifying fMRI runs for deletion due to excessive headmotion. Set to \code{0.25} by default}
}
\value{
outputs a .csv file containing the columns of fMRI runs and FD, and a bash script \code{del.sh}
}
\description{
Calculating mean FD from fmriprep outputs and generating bash script to remove fMRI runs with excessive headmotion
}
\details{
This function calculating mean RMSD from fmriprep outputs and generating bash script to remove fMRI runs with excessive headmotion.
}
\examples{
\dontrun{
headmotion.fmriprep()
}
}
