% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headmotion_XCP.R
\name{headmotion.XCP}
\alias{headmotion.XCP}
\title{headmotion.XCP}
\usage{
headmotion.XCP(filename = "FD.csv", threshold = 0.25)
}
\arguments{
\item{filename}{Filename of the output.csv file. Set to \code{FD.csv} by default}

\item{threshold}{mean FD threshold for identifying fMRI runs for deletion due to excessive headmotion. Set to \code{0.25} by default}
}
\value{
outputs a .csv file containing the columns of fMRI runs and FD, and a bash script \code{del.sh}
}
\description{
Calculating mean FD from XCP outputs and generating bash script to remove fMRI runs with excessive headmotion
}
\details{
This function calculating mean FD from XCP outputs and generating bash script to remove fMRI runs with excessive headmotion. This is used mainly on the XCP-d post-processed HCP fMRI data.
}
\examples{
\dontrun{
headmotion.XCP()
}
}
