% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyJSON.R
\name{modifyJSON}
\alias{modifyJSON}
\title{modifyJSON}
\usage{
modifyJSON(json_file, new_items)
}
\arguments{
\item{new_items}{one or multiple new items specified as a list object (e.g, \code{list(a=1, b=2)})}

\item{json_files}{filename suffix for one (e.g., \code{"_asl.json"})}
}
\description{
A tool to edit .JSON files
}
\details{
This function searches for .JSON files and adds or modify the one or multiple specified items
}
\examples{
\dontrun{
modifyJSON(json_files, new_items)
}
}
