% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordersub.R
\name{reorder_subcortical}
\alias{reorder_subcortical}
\title{reorder_subcortical}
\usage{
reorder_subcortical(data)
}
\arguments{
\item{data}{a vector of edge values with a length of 7021 or 23871}
}
\value{
a vector of edge values with the correct subcortical nodes ordering
}
\description{
to reorder subcortical nodes for visualization
}
\details{
The subcortical node elements in the previously extracted HCP matrices are not ordered correctly. Hence the visualizations (\code{vizHeatmap()} or \code{vizConnectogram()}) of the subcortical nodes generated from these HCP matrices are not accurate; Specifically the positioning of the subcortical nodes within the subcortical network is wrong, while the cortical nodes are unaffected. This function thus corrects the ordering of the subcortical nodes
}
\examples{

FC.vector=runif(23871, min = -1, max = 1)
FC.vector.reordered=reorder_subcortical(FC.vector)

}
